refactor(generator): split up python code into modules
