+++ spec-generator/example.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Example:
+    name: str
+    description: str = ''
+    text: str = ''
+    json: str = ''
+
